rename engines/bountyEngine.js to logic/bounties.js
