+++ source/buttons/bbaddcompleters.js
@@ -1,7 +1,7 @@
 const { ActionRowBuilder, UserSelectMenuBuilder, userMention, bold } = require('discord.js');
 const { ButtonWrapper } = require('../classes');
 const { SKIP_INTERACTION_HANDLING } = require('../constants');
-const { addCompleters } = require('../engines/bountyEngine');
+const { addCompleters } = require('../logic/bounties.js');
 const { listifyEN, commandMention } = require('../util/textUtil');
 
 const mainId = "bbaddcompleters";
